refactor(MilestoneService): Extract MilestoneReportService #1184
