+++ src/app/services/milestone/milestoneService.spec.ts
@@ -1,46 +1,37 @@
 import * as angular from 'angular';
-import { MilestoneService } from '../../assets/wise5/services/milestoneService';
+import { MilestoneService } from '../../../assets/wise5/services/milestoneService';
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { UpgradeModule } from '@angular/upgrade/static';
-import { AchievementService } from '../../assets/wise5/services/achievementService';
-import { ConfigService } from '../../assets/wise5/services/configService';
-import { ProjectService } from '../../assets/wise5/services/projectService';
-import { TeacherDataService } from '../../assets/wise5/services/teacherDataService';
-import { TeacherProjectService } from '../../assets/wise5/services/teacherProjectService';
-import { TeacherWebSocketService } from '../../assets/wise5/services/teacherWebSocketService';
-import { ClassroomStatusService } from '../../assets/wise5/services/classroomStatusService';
-import { CopyNodesService } from '../../assets/wise5/services/copyNodesService';
+import { AchievementService } from '../../../assets/wise5/services/achievementService';
+import { ConfigService } from '../../../assets/wise5/services/configService';
+import { ProjectService } from '../../../assets/wise5/services/projectService';
+import { TeacherDataService } from '../../../assets/wise5/services/teacherDataService';
+import { TeacherProjectService } from '../../../assets/wise5/services/teacherProjectService';
+import { TeacherWebSocketService } from '../../../assets/wise5/services/teacherWebSocketService';
+import { ClassroomStatusService } from '../../../assets/wise5/services/classroomStatusService';
+import { CopyNodesService } from '../../../assets/wise5/services/copyNodesService';
 import { MatDialogModule } from '@angular/material/dialog';
-import { StudentTeacherCommonServicesModule } from '../student-teacher-common-services.module';
-import aggregateAutoScoresSample from './sampleData/sample_aggregateAutoScores.json';
-import satisfyCriterionSample from './sampleData/sample_satisfyCriterion.json';
+import { StudentTeacherCommonServicesModule } from '../../student-teacher-common-services.module';
+import satisfyCriterionSample from '../sampleData/sample_satisfyCriterion.json';
+import { MilestoneReportService } from '../../../assets/wise5/services/milestoneReportService';
+import { createSatisfyCriteria, createScoreCounts } from './milestone-test-functions';
 
 let service: MilestoneService;
 let achievementService: AchievementService;
 let configService: ConfigService;
+let milestoneReportService: MilestoneReportService;
 let projectService: ProjectService;
 let teacherDataService: TeacherDataService;
-
-const reportSettingsCustomScoreValuesSample = {
-  customScoreValues: {
-    ki: [1, 2, 3, 4]
-  }
-};
 
 describe('MilestoneService', () => {
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule,
-        MatDialogModule,
-        StudentTeacherCommonServicesModule,
-        UpgradeModule
-      ],
+      imports: [HttpClientTestingModule, MatDialogModule, StudentTeacherCommonServicesModule],
       providers: [
         ClassroomStatusService,
         CopyNodesService,
         MilestoneService,
+        MilestoneReportService,
         TeacherDataService,
         TeacherProjectService,
         TeacherWebSocketService
@@ -49,6 +40,7 @@ describe('MilestoneService', () => {
     service = TestBed.inject(MilestoneService);
     achievementService = TestBed.inject(AchievementService);
     configService = TestBed.inject(ConfigService);
+    milestoneReportService = TestBed.inject(MilestoneReportService);
     projectService = TestBed.inject(ProjectService);
     teacherDataService = TestBed.inject(TeacherDataService);
   });
@@ -61,44 +53,9 @@ describe('MilestoneService', () => {
   insertMilestoneReport();
   getReferencedComponent();
   isCompletionReached();
-  generateReport();
-  chooseTemplate();
   getSatisfyCriteriaReferencedComponents();
-  adjustKIScore();
-  getKIScoreBounds();
-  addDataToAggregate();
-  setupAggregateSubScore();
-  getCustomScoreValueCounts();
-  getPossibleScoreValueCounts();
-  processMilestoneGraphsAndData();
   setReportAvailable();
 });
-
-function createScoreCounts(counts: any[]) {
-  const countsObject = {};
-  for (let c = 0; c < counts.length; c++) {
-    countsObject[c + 1] = counts[c];
-  }
-  return countsObject;
-}
-
-function createSatisfyCriteria(
-  nodeId: string,
-  componentId: string,
-  targetVariable: string = null,
-  func: string = null,
-  value: number = null,
-  percentThreshold: number = null
-) {
-  return {
-    nodeId: nodeId,
-    componentId: componentId,
-    targetVariable: targetVariable,
-    function: func,
-    value: value,
-    percentThreshold: percentThreshold
-  };
-}
 
 function getProjectMilestones() {
   describe('getProjectMilestones()', () => {
@@ -354,7 +311,9 @@ function insertMilestoneReport() {
           scoreCount: 50
         }
       };
-      spyOn(service, 'calculateAggregateAutoScores').and.returnValue(aggregateAutoScores);
+      spyOn(milestoneReportService, 'calculateAggregateAutoScores').and.returnValue(
+        aggregateAutoScores
+      );
       service.insertMilestoneReport(milestone);
       expect(milestone.isReportAvailable).toEqual(true);
       expect(milestone.generatedReport).toEqual(content);
@@ -437,67 +396,6 @@ function isCompletionReached() {
   });
 }
 
-function generateReport() {
-  it('should generate report', () => {
-    const content = 'template1Content';
-    const projectAchievement = {
-      report: {
-        locations: [
-          {
-            nodeId: 'node1',
-            componentId: 'component1'
-          }
-        ],
-        templates: [
-          {
-            id: 'template1',
-            satisfyConditional: 'any',
-            satisfyCriteria: [
-              createSatisfyCriteria(
-                'node1',
-                'component1',
-                'ki',
-                'percentOfScoresLessThanOrEqualTo',
-                3,
-                50
-              )
-            ],
-            content: content
-          }
-        ]
-      }
-    };
-    const aggregateAutoScores = {
-      ki: {
-        counts: createScoreCounts([10, 10, 10, 10, 10]),
-        scoreCount: 50
-      }
-    };
-    spyOn(service, 'calculateAggregateAutoScores').and.returnValue(aggregateAutoScores);
-    const report = service.generateReport(projectAchievement);
-    expect(report.content).toEqual(content);
-  });
-}
-
-function chooseTemplate() {
-  describe('chooseTemplate()', () => {
-    it('should choose template', () => {
-      const template1 = {
-        id: 'template-1'
-      };
-      const template2 = {
-        id: 'template-2'
-      };
-      const templates = [template1, template2];
-      const aggregateAutoScores = [];
-      spyOn(service, 'isTemplateMatch').and.callFake((template, aggregateAutoScores) => {
-        return template.id === 'template-2';
-      });
-      expect(service.chooseTemplate(templates, aggregateAutoScores)).toEqual(template2);
-    });
-  });
-}
-
 function getSatisfyCriteriaReferencedComponents() {
   describe('getSatisfyCriteriaReferencedComponents()', () => {
     it('should return referenced components', () => {
@@ -522,155 +420,6 @@ function getSatisfyCriteriaReferencedComponents() {
           componentId: 'xfns1g7pga'
         }
       });
-    });
-  });
-}
-
-// TODO: finish
-function calculateAggregateAutoScores() {
-  describe('calculateAggregateAutoScores()', () => {
-    it('should return the aggregate auto scores', () => {});
-  });
-}
-
-function adjustKIScore() {
-  describe('adjustKIScore()', () => {
-    it('should return the adjusted KI score', () => {
-      const value = 5;
-      expect(service.adjustKIScore(value, reportSettingsCustomScoreValuesSample)).toEqual(4);
-    });
-  });
-}
-
-function getKIScoreBounds() {
-  describe('getKIScoreBounds()', () => {
-    it('should return the KI score bounds', () => {
-      expect(service.getKIScoreBounds(reportSettingsCustomScoreValuesSample)).toEqual({
-        min: 1,
-        max: 4
-      });
-    });
-  });
-}
-
-function addDataToAggregate() {
-  describe('addDataToAggregate()', () => {
-    it('should add annotation to the aggregate scores and return aggregate', () => {
-      const annotation = {
-        data: {
-          scores: [
-            {
-              id: 'ki',
-              score: 3
-            }
-          ]
-        }
-      };
-      const aggregateAutoScore = angular.copy(aggregateAutoScoresSample)[0].aggregateAutoScore;
-      const result = service.addDataToAggregate(
-        aggregateAutoScore,
-        annotation,
-        reportSettingsCustomScoreValuesSample
-      );
-      expect(result).toEqual({
-        ki: {
-          counts: { 1: 2, 2: 0, 3: 2, 4: 0, 5: 0 },
-          scoreSum: 8,
-          scoreCount: 4,
-          average: 2
-        }
-      });
-    });
-  });
-}
-
-function setupAggregateSubScore() {
-  describe('setupAggregateSubScore()', () => {
-    it('should setup aggregate sub score', () => {
-      const subScoreId = 'ki';
-      const reportSettings = {};
-      const counts = service.setupAggregateSubScore(subScoreId, reportSettings);
-      expect(counts.scoreSum).toEqual(0);
-      expect(counts.scoreCount).toEqual(0);
-      expect(counts.counts).toEqual(createScoreCounts([0, 0, 0, 0, 0]));
-      expect(counts.average).toEqual(0);
-    });
-    it('should setup aggregate sub score with custom score values', () => {
-      const subScoreId = 'ki';
-      const reportSettings = {
-        customScoreValues: {
-          ki: [0, 1, 2]
-        }
-      };
-      const counts = service.setupAggregateSubScore(subScoreId, reportSettings);
-      expect(counts.scoreSum).toEqual(0);
-      expect(counts.scoreCount).toEqual(0);
-      expect(counts.counts).toEqual({ 0: 0, 1: 0, 2: 0 });
-      expect(counts.average).toEqual(0);
-    });
-  });
-}
-
-function getCustomScoreValueCounts() {
-  describe('getCustomScoreValueCounts()', () => {
-    it('should get custom score value counts', () => {
-      const scoreValues = service.getCustomScoreValueCounts([0, 1, 2]);
-      expect(Object.entries(scoreValues).length).toEqual(3);
-      expect(scoreValues[0]).toEqual(0);
-      expect(scoreValues[1]).toEqual(0);
-      expect(scoreValues[2]).toEqual(0);
-    });
-  });
-}
-
-function getPossibleScoreValueCounts() {
-  describe('getPossibleScoreValueCounts()', () => {
-    it('should get possible score value counts for ki', () => {
-      const scoreValues = service.getPossibleScoreValueCounts('ki');
-      expect(Object.entries(scoreValues).length).toEqual(5);
-      expect(scoreValues[1]).toEqual(0);
-      expect(scoreValues[2]).toEqual(0);
-      expect(scoreValues[3]).toEqual(0);
-      expect(scoreValues[4]).toEqual(0);
-      expect(scoreValues[5]).toEqual(0);
-    });
-    it('should get possible score value counts for not ki', () => {
-      const scoreValues = service.getPossibleScoreValueCounts('science');
-      expect(Object.entries(scoreValues).length).toEqual(3);
-      expect(scoreValues[1]).toEqual(0);
-      expect(scoreValues[2]).toEqual(0);
-      expect(scoreValues[3]).toEqual(0);
-    });
-  });
-}
-
-function processMilestoneGraphsAndData() {
-  describe('processMilestoneGraphsAndData()', () => {
-    it('should process milestone report graph', () => {
-      let content = '<milestone-report-graph id="ki"></milestone-report-graph>';
-      const componentAggregateAutoScores = [
-        {
-          nodeId: 'node1',
-          componentId: 'component1',
-          aggregateAutoScore: {
-            ki: {
-              scoreSum: 4,
-              scoreCount: 2,
-              average: 2,
-              counts: createScoreCounts([1, 0, 1, 0, 0])
-            }
-          }
-        }
-      ];
-      content = service.processMilestoneGraphsAndData(content, componentAggregateAutoScores);
-      expect(
-        content.includes(
-          `data="[{` +
-            `'scoreSum':4,'scoreCount':2,'average':2,'counts':{'1':1,'2':0,'3':1,'4':0,'5':0},` +
-            `'nodeId':'node1','componentId':'component1'` +
-            `}]"`
-        )
-      ).toEqual(true);
     });
   });
 }
