+++ src/assets/wise5/classroomMonitor/classroom-monitor-testing.module.ts
@@ -11,6 +11,7 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MilestoneService } from '../services/milestoneService';
 import { TeacherPeerGroupService } from '../services/teacherPeerGroupService';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MilestoneReportService } from '../services/milestoneReportService';
 
 @NgModule({
   imports: [
@@ -24,6 +25,7 @@ import { MatSnackBarModule } from '@angular/material/snack-bar';
   providers: [
     ClassroomStatusService,
     MilestoneService,
+    MilestoneReportService,
     TeacherDataService,
     TeacherPeerGroupService,
     TeacherProjectService,
